Allow to disable qs footer warnings [1/2]
When using apps like AdGuard that create a vpn, a warning is displayed
in the QS panel footer, so the user knows the device net is being monitored
by an external app.
The warning takes space and the user might want to avoid it if he already
knows that the app is running.

Signed-off-by: spezi77 <spezi77@gmx.com>
Signed-off-by: Funny-Bunny <dragos281993@gmail.com>